+++ src/components/internal/index.ts
@@ -1,1 +1,2 @@
 export { default as StoryPanel } from './StoryPanel.vue';
+export { default as InputWithContentWrapper } from './InputWithContentWrapper.vue';